refactor(streams): extract subjectCovers into subject-utils
