+++ src/server/infrastructure/stream.provider.ts
@@ -33,25 +33,10 @@ import { mapProvisioningError, type ProvisioningErrorContext } from './provision
 import { formatProvisioningSummary, type StreamReservation } from './provisioning-summary';
 import { compareStreamConfig, type StreamConfigDiffResult } from './stream-config-diff';
 import { StreamMigration } from './stream-migration';
+import { subjectCovers } from './subject-utils';
 
 /** A `StreamKind` or the `'dlq'` label, used for reservation/error provenance. */
 type ReservationKind = StreamKind | 'dlq';
-
-/** True when `broad` matches everything `narrow` matches. Identical entries return false. */
-const subjectCovers = (broad: string, narrow: string): boolean => {
-  if (broad === narrow) return false;
-
-  const broadTokens = broad.split('.');
-  const narrowTokens = narrow.split('.');
-
-  for (let i = 0; i < broadTokens.length; i += 1) {
-    if (broadTokens[i] === '>') return i < narrowTokens.length;
-    if (i >= narrowTokens.length || narrowTokens[i] === '>') return false;
-    if (broadTokens[i] !== '*' && broadTokens[i] !== narrowTokens[i]) return false;
-  }
-
-  return broadTokens.length === narrowTokens.length;
-};
 
 /**
  * Manages JetStream stream lifecycle: creation, updates, and idempotent ensures.
